task 22

+++ Seminar/c#_21/Program.cs
@@ -1,29 +1,29 @@
 ﻿Console.Clear();
-int [] coord1 = new int [2];
+double [] coord1 = new double [2];
 int i = 0;
 Console.Write("Первая точка имеет координаты (X,Y): ");
 while (i<coord1.Length )
 {
-    coord1[i] = new Random().Next(-50,51);
+    coord1[i] = new  Random().Next(-50,51);
     Console.Write(coord1[i] + " ");
     i++;
 }
-int [] coord2 = new int [2];
+double [] coord2 = new double [2];
 i = 0;
 Console.Write("\nВторая точка имеет координаты (X,Y): ");
 while (i<coord2.Length )
 {
-    coord2[i] = new Random().Next(-50,51);
+    coord2[i] = new  Random().Next(-50,51);
     Console.Write(coord2[i] + " ");
     i++;
 }
 Console.WriteLine();
 
 i = 0;
-int sum =0;
+double sum =0;
 while (i<coord1.Length)
 {
-    int x = Math.Abs (coord1[i] - coord2[i]);
+    double x = Math.Abs (coord1[i] - coord2[i]);
     sum = sum + x*x;
     i++;
 }
